+++ src/omnivoxel/client/game/graphics/opengl/mesh/util/PriorityUtils.java
@@ -8,10 +8,9 @@ public final class PriorityUtils {
     private static Camera camera;
 
     public static double getPriority(Position3D position) {
-        // The closer a chunk is the higher the priority
-        int distanceX = position.x() * ConstantGameSettings.CHUNK_WIDTH - Math.round(camera.getX());
-        int distanceY = position.y() * ConstantGameSettings.CHUNK_HEIGHT - Math.round(camera.getY());
-        int distanceZ = position.z() * ConstantGameSettings.CHUNK_LENGTH - Math.round(camera.getZ());
+        long distanceX = (long) position.x() * ConstantGameSettings.CHUNK_WIDTH - Math.round(camera.getX());
+        long distanceY = (long) position.y() * ConstantGameSettings.CHUNK_HEIGHT - Math.round(camera.getY());
+        long distanceZ = (long) position.z() * ConstantGameSettings.CHUNK_LENGTH - Math.round(camera.getZ());
         return (distanceX * distanceX + distanceY * distanceY + distanceZ * distanceZ) / 1000.0;
     }
