+++ src/omnivoxel/client/network/request/PlayerUpdateRequest.java
@@ -1,6 +1,6 @@
 package omnivoxel.client.network.request;
 
-public record PlayerUpdateRequest(float x, float y, float z, float pitch, float yaw) implements Request {
+public record PlayerUpdateRequest(double x, double y, double z, double pitch, double yaw) implements Request {
     @Override
     public RequestType getType() {
         return RequestType.PLAYER_UPDATE;
